requirements.txt
vectorbt==0.26.3
pandas>=2.2
numpy>=1.26
numba>=0.59
ccxt>=4.3
python-dotenv>=1.0
PyYAML>=6.0
loguru>=0.7
matplotlib>=3.8
